% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgen.R
\docType{class}
\name{ClassSupport-class}
\alias{ClassSupport-class}
\title{class definition support, automates creation of getters and setters}
\description{
class definition support, automates creation of getters and setters
}
