% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgen.R
\name{clgen}
\alias{clgen}
\title{write the code for an S4 class with setters and getters}
\usage{
clgen(
  clname = "Turtle",
  protolist = list(location = c(0, 0), orientation = 0, path = matrix(c(0, 0), ncol =
    2)),
  check_init = FALSE
)
}
\arguments{
\item{clname}{character(1) class name}

\item{protolist}{named list() of prototypical values for slots}

\item{check_init}{logical(1) if TRUE, try to construct instance with prototype before building code}
}
\description{
write the code for an S4 class with setters and getters
}
\note{
The example is inspired by \href{https://stuartlee.org/2019/07/09/s4-short-guide/}{Stuart Lee's blog post}
}
\examples{
dm = clgen()
dm
dumpcs(dm)
}
