% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defdMethods.R
\name{defdMethods}
\alias{defdMethods}
\title{defdMethods enumerates methods defined for a class in a package -- NEEDS WORK}
\usage{
defdMethods(cl, pkg = paste0("package:", cl))
}
\arguments{
\item{cl}{character(1) name of class}

\item{pkg}{character(1) name of package where class is defined, defaults to 'package:[classname]' which works for major classes like SummarizedExperiment}
}
\description{
defdMethods enumerates methods defined for a class in a package -- NEEDS WORK
}
\note{
This is an early attack on the problem of enumerating
defined methods for a class.  As of 5 Jan 2019 this function
lists only one signature per method and fails on oddly
names signature components like '_data'.
}
\examples{
require("SingleCellExperiment")
scem = defdMethods("SingleCellExperiment")
scem
scem[[2]][[1]]
}
