% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clgen.R
\name{dumpcs,ClassSupport-method}
\alias{dumpcs,ClassSupport-method}
\title{output definition, getters, setters}
\usage{
\S4method{dumpcs}{ClassSupport}(csupp, con = stdout())
}
\description{
output definition, getters, setters
}
